+++ components/missions/Preview.vue
@@ -1,16 +1,17 @@
 <template>
   <n-link :to="`/mission/${mission.id}`">
-    <div class="card">
+    <v-card>
       <v-picture :url="mission.thumb_url" />
       <h3>{{mission.name}}</h3>
       <v-status :status="mission.status" />
       <v-due-date :due-date="mission.due_date" />
       <v-progress :status="mission.status" :progress="mission.enrollment || null" />
-    </div>
+    </v-card>
   </n-link>
 </template>
 
 <script>
+import vCard from "./shared/Card";
 import vPicture from "./shared/Picture";
 import vStatus from "./shared/Status";
 import vDueDate from "./shared/DueDate";
@@ -23,6 +24,7 @@ export default {
     }
   },
   components: {
+    vCard,
     vPicture,
     vStatus,
     vDueDate,
@@ -32,27 +34,6 @@ export default {
 </script>
 
 <style lang="sass" scoped>
-.card
-  width: 340px
-  background-color: rgba(white, .05)
-  display: flex
-  flex-direction: column
-  align-items: center
-  margin: 20px
-  padding: 20px 30px 0
-  border-radius: 8px
-  margin-top: 60px
-  color: white
-  animation: fade .5s
-
-  &:hover
-    transform: translateY(-5px)
-
-  h3
-    margin: 10px 0
-    font-weight: 700
-    font-size: 24px
-
 a
   text-decoration: none
